+++ KillOnSight-midnight/Locale/daDK.lua
@@ -66,6 +66,8 @@ L.UI_NEARBY_MINIMAL = "Ultra-minimalt nær-vindue"
 L.UI_NEARBY_ROWFADE = "Fade-timere pr. række"
 L.UI_NEARBY_ROWICONS = "Rækkeikoner (klasse/kranie)"
 L.UI_NEARBY_SCALE = "Skala for nær-vindue"
+L.UI_NEARBY_NAME_FONT = "Skrifttype for navne i nærheden"
+L.UI_NEARBY_NAME_SIZE = "Skriftstørrelse for navne i nærheden"
 L.UI_OPTIONS = "Options"
 
 L.UI_OPTIONS_TITLE = "Indstillinger"