+++ KillOnSight-midnight/Locale/jaJP.lua
@@ -68,6 +68,8 @@ L.UI_NEARBY_MINIMAL = "超ミニマル付近ウィンドウ"
 L.UI_NEARBY_ROWFADE = "行ごとのフェードタイマー"
 L.UI_NEARBY_ROWICONS = "行アイコン (クラス/髑髏)"
 L.UI_NEARBY_SCALE = "付近ウィンドウの倍率"
+L.UI_NEARBY_NAME_FONT = "近くの名前のフォント"
+L.UI_NEARBY_NAME_SIZE = "近くの名前のサイズ"
 L.UI_OPTIONS = "オプション"
 
 L.UI_OPTIONS_TITLE = "オプション"